requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9  # For PostgreSQL
python-dotenv==1.0.0
yfinance==0.2.55
drf-nested-routers
